+++ src/app/Types.ts
@@ -1,5 +1,0 @@
-export class Types {
-    public static readonly TENANT_RESOLVER = 'common-web.TenantResolver'
-    public static readonly WEBSERVER_ADDON = 'common-web.ExpressServerAddOn'
-    public static readonly AUTH_ADDON = 'common-web.AuthAddOn'
-}