+++ src/app/constants/Types.ts
@@ -1,0 +1,4 @@
+export class Types {
+    public static readonly TENANT_RESOLVER = 'web.TenantResolver'
+    public static readonly WEBSERVER_ADDON = 'web.ExpressServerAddOn'
+}