+++ src/app/index.ts
@@ -2,16 +2,16 @@
 import decoratorObj = require('./decorators/index')
 export const decorators = decoratorObj.decorators
 
-export * from './constants/AuthConstant'
 export * from './constants/MetaData'
-export { IActionFilter, IActionErrorHandler, FilterPriority } from './decorators/filter'
-export * from './filters/AuthorizeFilter'
+export { IActionFilter, IActionErrorHandler, FilterPriority,
+    addFilterToTarget, pushFilterToArray } from './decorators/filter'
+export * from './interfaces'
+export * from './filters/ActionFilterBase'
 export * from './filters/ErrorHandlerFilter'
 export * from './filters/ModelFilter'
 export * from './filters/TenantResolverFilter'
-export * from './AuthAddOn'
 export * from './ExpressServerAddOn'
 export * from './RestControllerBase'
 export * from './register-addon'
-export * from './Types'
+export * from './constants/Types'
 export * from './WebContext'