+++ src/test/shared/default-controller.ts
@@ -1,40 +1,39 @@
-import * as express from 'express'
-import { decorators as dec } from '../../app'
+import { decorators as dec, Request, Response } from '../../app'
 
 @dec.controller()
 class DefaultController {
     @dec.GET()
-    public doGet(req: express.Request, res: express.Response): void {
+    public doGet(req: Request, res: Response): void {
         res.send('DefaultController.doGet')
     }
 
     @dec.POST()
-    public doPost(req: express.Request, res: express.Response): void {
+    public doPost(req: Request, res: Response): void {
         res.send('DefaultController.doPost')
     }
 
     @dec.PATCH()
-    public doPatch(req: express.Request, res: express.Response): void {
+    public doPatch(req: Request, res: Response): void {
         res.send('DefaultController.doPatch')
     }
 
     @dec.PUT()
-    public doPut(req: express.Request, res: express.Response): void {
+    public doPut(req: Request, res: Response): void {
         res.send('DefaultController.doPut')
     }
 
     @dec.DELETE()
-    public doDelete(req: express.Request, res: express.Response): void {
+    public doDelete(req: Request, res: Response): void {
         res.send('DefaultController.doDelete')
     }
 
     @dec.HEAD()
-    public doHead(req: express.Request, res: express.Response): void {
+    public doHead(req: Request, res: Response): void {
         res.send('DefaultController.doHead')
     }
 
     @dec.OPTIONS()
-    public doOptions(req: express.Request, res: express.Response): void {
+    public doOptions(req: Request, res: Response): void {
         res.send('DefaultController.doOptions')
     }
 
@@ -45,19 +44,19 @@ class DefaultController {
     @dec.DELETE()
     @dec.HEAD()
     @dec.OPTIONS()
-    public doMany(req: express.Request, res: express.Response): void {
+    public doMany(req: Request, res: Response): void {
         res.send('DefaultController.doMany')
     }
 
     @dec.ALL()
-    public doAll(req: express.Request, res: express.Response): void {
+    public doAll(req: Request, res: Response): void {
         res.send('DefaultController.doAll')
     }
 
     /**
      * Intentionally no action decorator
      */
-    public doSecret(req: express.Request, res: express.Response): void {
+    public doSecret(req: Request, res: Response): void {
         res.send('DefaultController.doSecret')
     }
 }